fix app crashes on startup and add database folder validation
- Fixed SIGABRT crash caused by calling `ReusableFunc.showAlert` from a background thread by wrapping the presentation logic in `DispatchQueue.main.async`.
- Fixed SIGTRAP crash on iOS where `DatabaseManager.handleSetupError` called `fatalError` when the database folder was missing. It now returns gracefully without forcefully terminating the app.
- Fixed deadlock (SIGTRAP via `_dispatch_once_wait`) during `HistoryViewModel` initialization by deferring the CloudKit history upload (`CloudKitSyncManager.shared.uploadHistory`) into a `Task`, avoiding recursive lock on the shared instance.
- Implemented strict database folder validation (`validateDatabaseFolder`) to ensure the selected folder contains `Files/main.sqlite` and `Files/special.sqlite` with the correct schema before accepting the path.
- Extracted `SettingsViewModel` from `SettingsView.swift` into its own file (`Source/Managers/ViewModels/SettingsViewModel.swift`) for cleaner architecture and maintainability.
- Updated `SettingsActions.selectLibraryFolder` to accept and execute the new validation closure.
